feat: fetch all episodes on endpoint /episodes
